==> scripts/02_get_infos_gaemas.ipynb <==
<br>

# GAEMAs


import pandas as pd
from open_geodata import geo

from sp_mpsp_divadmin.paths import input_path_tab, output_path_tabs

<br>

# Relatório


<br>

## Nomes dos Municípios

Lê o arquivo csv com o nome dos municípios


#geo.get_dataset_names()

df_mun = geo.load_dataset('tab.sp.tab_municipio_nome')

df_mun.info()
df_mun.head()

<br>

## IDs vinculando Municípios com GAEMAs


# Dataframe
df = pd.read_csv(
    input_path_tab / 'tab_ids_municipio.csv',
    usecols=['id_municipio', 'id_gaema'],
)

# Results
df.info()
df.head()

# Merge Municípios
df = pd.merge(
    df_mun,
    df,
    how='left',
    left_on='id_municipio',
    right_on='id_municipio',
)

# Results
df.info()
df.head()

<br>

## Município GAEMAs


# Dataframe GAEMAs
df_gaema = pd.read_csv(input_path_tab / 'tab_gaema.csv')
df_gaema = df_gaema.add_prefix('gaema_')

# Results
df_gaema.info()
df_gaema.head()

# Merge GAEMAs
df = pd.merge(
    df,
    df_gaema,
    how='left',
    left_on='id_gaema',
    right_on='gaema_id',
)

# Drop
df.dropna(subset=['gaema_id'], inplace=True)
df.drop(['gaema_id'], axis=1, inplace=True, errors='ignore')

# Save
df.to_csv(
    output_path_tabs / 'tab_municipio_gaema.csv',
    index=False,
)

# Results
df.info()
df.head()
