==> scripts/03_get_infos_gaeco.ipynb <==
<br>

# GAECOs


import geopandas as gpd
import pandas as pd
from open_geodata import geo

from sp_mpsp_divadmin.paths import input_path_tab, output_path_tabs, output_path_gpkg

<br>

## Nomes dos Municípios

Lê o arquivo csv com o nome dos municípios


df_mun = geo.load_dataset('tab.sp.tab_municipio_nome')

df_mun.info()
df_mun.head()

<br>

## IDs vinculando Municípios com Redes Protetivas


# Dataframe
df = pd.read_csv(
    input_path_tab / 'tab_ids_municipio.csv',
    usecols=['id_municipio', 'id_gaeco_mun'],
)
# Results
df.info()
df.head()

# Merge Municípios
df = pd.merge(
    df_mun,
    df,
    how='left',
    left_on='id_municipio',
    right_on='id_municipio',
)

# Results
df.info()
df.head()

<br>

## Município GAECO


# Dataframe Rede Protetiva
df_gaeco = pd.read_csv(input_path_tab / 'tab_gaeco.csv')
df_gaeco = df_gaeco.add_prefix('gaeco_')

# Results
df_gaeco.info()
df_gaeco.head()

# Merge Rede Protetiva
df = pd.merge(
    df,
    df_gaeco,
    how='left',
    left_on='id_gaeco_mun',
    right_on='gaeco_id',
)

# Drop
df.dropna(subset=['gaeco_id'], inplace=True)
df.drop(['gaeco_id'], axis=1, inplace=True, errors='ignore')

# Save
df.to_csv(
    output_path_tabs / 'tab_municipio_gaeco.csv',
    index=False,
)

# Results
df.info()
df.head()
