==> scripts/04_get_infos_macrorregioes.ipynb <==
<br>

# Macroregiões

import pandas as pd
from open_geodata import geo

from sp_mpsp_divadmin.paths import input_path_tab, output_path_tabs

<br>

## Read Data

### Nomes dos Municípios

Lê o arquivo csv com o nome dos municípios

df_mun = geo.load_dataset('tab.sp.tab_municipio_nome')

df_mun.info()
df_mun.head()

<br>

## IDs vinculando Municípios com Áreas Regionais

# Dataframe
df_ar = pd.read_csv(
    input_path_tab / 'tab_ids_municipio.csv',
    usecols=['id_municipio', 'id_ar'],
)

# Results
df_ar.info()
df_ar.head()

# Dataframe
df_ar_macro = pd.read_csv(
    input_path_tab / 'tab_ids_macrorregioes.csv',
)

# Results
df_ar_macro.info()
df_ar_macro.head()

# Dataframe GAEMAs
df_macro = pd.read_csv(input_path_tab / 'tab_macrorregioes.csv')
df_macro = df_macro.add_prefix('macrorregiao_')

# Results
df_macro.info()
df_macro.head()

<br>

## Município - AR

# Merge Municípios
df = pd.merge(
    df_ar,
    df_mun,
    how='left',
    left_on='id_municipio',
    right_on='id_municipio',
)

# Results
df.info()
df.head()

<br>

## IDs vinculando Áreas Regionais com Macroregiões

# Merge Municípios
df = pd.merge(
    df,
    df_ar_macro,
    how='left',
    left_on='id_ar',
    right_on='id_ar',
)

# Results
df.info()
df.head()

<br>

## Município Macroregiões

# Merge GAEMAs
df = pd.merge(
    df,
    df_macro,
    how='left',
    left_on='id_macrorregiao',
    right_on='macrorregiao_id',
)

# Results
df.info()
df.head()

# Drop
df.drop(['id_ar', 'macrorregiao_id'], axis=1, inplace=True, errors='ignore')

# Results
df.info()
df.head()

# Save
df.to_csv(
    output_path_tabs / 'tab_municipio_macrorregioes.csv',
    index=False,
)

# Results
df.info()
df.head()
