==> scripts/11_create_shps.ipynb <==
<br>

# Create _geojson_


import geopandas as gpd
import pandas as pd
from open_geodata import geo
from tqdm.notebook import tqdm, trange

from sp_mpsp_divadmin.paths import (
    output_path_geo,
    output_path_gpkg,
    output_path_tabs,
)

gdf_sp = geo.load_dataset('geo.sp.sp_250k_wgs84')
gdf_sp.drop(['municipio_nome'], axis=1, inplace=True)
gdf_sp.head()

df_mun = geo.load_dataset('tab.sp.tab_municipio_nome')

df_mun.info()
df_mun.head()

gdf_sp = gdf_sp.merge(
    df_mun, how='left', left_on='id_municipio', right_on='id_municipio'
)
gdf_sp.head(2)

<br>

## GAEMA


# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    output_path_tabs / 'tab_municipio_gaema.csv',
)

# Drop Column
df = df.drop(['municipio_nome'], axis=1)

# Results
df.info()
df.head(2)

# Merge
gdf_gaema = gdf_sp.merge(
    df,
    on='id_municipio',
    how='right',
)

# Results
gdf_gaema.info()
gdf_gaema.head(2)

# Save geojson
gdf_gaema.to_file(
    output_path_gpkg / 'sp_mpsp_gaema.gpkg',
    layer='sp_mpsp_gaema',
    driver='GPKG',
)

gdf_gaema.to_file(
    output_path_gpkg / 'sp_mpsp.gpkg',
    layer='gaema',
    driver='GPKG',
)

# Results
gdf_gaema.head()

<br>

## Rede Protetiva


# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    output_path_tabs / 'tab_municipio_rp.csv',
)

# Drop Column
df = df.drop(['municipio_nome'], axis=1, errors='ignore')

# Results
df.info()
df.head(2)

# Merge
gdf_rp = gdf_sp.merge(
    df,
    on='id_municipio',
    how='right',
)

gdf_rp.head(2)

# Save geojson
gdf_rp.to_file(
    output_path_gpkg / 'sp_mpsp_rp.gpkg',
    layer='sp_mpsp_rp',
    driver='GPKG',
)

gdf_rp.to_file(
    output_path_gpkg / 'sp_mpsp.gpkg',
    layer='rede_protetiva',
    driver='GPKG',
)

# Results
gdf_rp.head(2)

<br>

## Área Regional


# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    output_path_tabs / 'tab_municipio_ar.csv',
)

# Drop Column
df = df.drop(['municipio_nome'], axis=1, errors='ignore')

# Results
df.info()
df.head(2)

# Merge
gdf_ar = gdf_sp.merge(
    df,
    on='id_municipio',
    how='left',
)

# Save geojson
gdf_ar.to_file(
    output_path_gpkg / 'sp_mpsp_ar.gpkg',
    layer='sp_mpsp_ar',
    driver='GPKG',
)

gdf_ar.to_file(
    output_path_gpkg / 'sp_mpsp.gpkg',
    layer='area_regional',
    driver='GPKG',
)

# Results
gdf_ar.info()
gdf_ar.head(2)

<br>

## Macrorregiões


# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    output_path_tabs / 'tab_municipio_macrorregioes.csv',
)
# Drop Column
df = df.drop(['municipio_nome'], axis=1, errors='ignore')

# Results
df.info()
df.head(2)

# Merge
gdf_macroregioes = gdf_sp.merge(
    df,
    on='id_municipio',
    how='left',
)

# Save geojson
gdf_macroregioes.to_file(
    output_path_gpkg / 'sp_mpsp_macrorregioes.gpkg',
    layer='sp_mpsp_macrorregioes',
    driver='GPKG',
)

gdf_macroregioes.to_file(
    output_path_gpkg / 'sp_mpsp.gpkg',
    layer='macrorregioes',
    driver='GPKG',
)

# Results
gdf_macroregioes.info()
gdf_macroregioes.head(2)

<br>

## PJ Regional

# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    output_path_tabs / 'tab_municipio_pjreg.csv',
)
# Drop Column
df = df.drop(['municipio_nome'], axis=1, errors='ignore')

# Results
df.info()
df.head(2)

# Merge
gdf_pjreg = gdf_sp.merge(
    df,
    on='id_municipio',
    how='left',
)

# Save geojson
gdf_pjreg.to_file(
    output_path_gpkg / 'sp_mpsp_pjreg.gpkg',
    layer='sp_mpsp_pjreg',
    driver='GPKG',
)

gdf_pjreg.to_file(
    output_path_gpkg / 'sp_mpsp.gpkg',
    layer='pjreg',
    driver='GPKG',
)

# Results
gdf_pjreg.info()
gdf_pjreg.head(2)

<br>

## GAECOs

# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    output_path_tabs / 'tab_municipio_gaeco.csv',
)
# Drop Column
df = df.drop(['municipio_nome'], axis=1, errors='ignore')

# Results
df.info()
df.head(2)

# Merge
gdf_gaeco = gdf_sp.merge(
    df,
    on='id_municipio',
    how='left',
)

# Save geojson
gdf_gaeco.to_file(
    output_path_gpkg / 'sp_mpsp_gaeco.gpkg',
    layer='sp_mpsp_gaeco',
    driver='GPKG',
)

gdf_gaeco.to_file(
    output_path_gpkg / 'sp_mpsp.gpkg',
    layer='gaeco',
    driver='GPKG',
)

# Results
gdf_gaeco.info()
gdf_gaeco.head(2)
