==> scripts/13_convert_7zip.ipynb <==
<br>

# Convert para 7 zip

import open_geodata

from sp_mpsp_divadmin.paths import (
    output_path_geo,
    output_path_gpkg,
    output_path_tabs,
)

open_geodata.converts.files.convert_to_7zip(input_path=output_path_gpkg, output_path=output_path_geo)
