==> scripts/21_get_infos_rps.ipynb <==
<br>

# Redes Protetivas

import os

import pandas as pd
from functions import *
from paths import *

# import re
# import time
# import random
# import tabula
# import requests
# import numpy as np
# import geopandas as gpd
# from bs4 import BeautifulSoup

<br>

# Relatório

<br>

## Nomes dos Municípios

Lê o arquivo csv com o nome dos municípios

df_mun.info()
df_mun.head()

<br>

## IDs vinculando Municípios com Redes Protetivas

# Dataframe
df = pd.read_csv(
    os.path.join(input_path_tab, 'tab_ids_municipio.csv'),
    usecols=['id_municipio', 'id_rp'],
)
# Results
df.info()
df.head()

# Merge Municípios
df = pd.merge(
    df_mun,
    df,    
    how='left',
    left_on='id_municipio',
    right_on='id_municipio',
)

# Results
df.info()
df.head()

<br>

## Município Rede Protetiva

# Dataframe Rede Protetiva
df_rp = pd.read_csv(os.path.join(input_path_tab, 'tab_redeprotetiva.csv'))
df_rp = df_rp.add_prefix('rp_')

# Results
df_rp.info()
df_rp.head()

# Merge Rede Protetiva
df = pd.merge(
    df,
    df_rp,
    how='left',
    left_on='id_rp',
    right_on='rp_id',
)

# Drop
df.dropna(subset=['rp_id'], inplace=True)
df.drop(['rp_id'], axis=1, inplace=True, errors='ignore')

# Save
df.to_csv(
    os.path.join(output_path_tabs, 'tab_municipio_rp.csv'),
    index=False,
)

# Results
df.info()
df.head()
