==> scripts/22_get_infos_ar.ipynb <==
<br>

# Áreas Regionais

#!pip3 install open-geodata --upgrade

import os

import pandas as pd
from functions import df_mun
from open_geodata import geo
from paths import *

<br>

## Nomes dos Municípios

Lê o arquivo csv com o nome dos municípios

df_mun.info()
df_mun.head()

<br>

## IDs vinculando Municípios com Áreas Regionais

# Dataframe
df = pd.read_csv(
    os.path.join(input_path_tab, 'tab_ids_municipio.csv'),
    usecols=['id_municipio', 'id_ar'],
)

# Results
df.info()
df.head()

# Merge Municípios
df = pd.merge(
    df_mun,
    df,    
    how='left',
    left_on='id_municipio',
    right_on='id_municipio',
)

# Results
df.info()
df.head()

<br>

## Município Áreas Regionais

# Dataframe GAEMAs
df_ar = pd.read_csv(os.path.join(input_path_tab, 'tab_arearegional.csv'))
df_ar = df_ar.add_prefix('ar_')

# Results
df_ar.info()
df_ar.head()

# Merge GAEMAs
df = pd.merge(
    df,
    df_ar,
    how='left',
    left_on='id_ar',
    right_on='ar_id',
)

# Drop
df.drop(['ar_id'], axis=1, inplace=True, errors='ignore')

# Save
df.to_csv(
    os.path.join(output_path_tabs, 'tab_municipio_ar.csv'),
    index=False,
)

# Results
df.info()
df.head()
