==> scripts/31_create_shps.ipynb <==
<br>

# Create *geojson*

import os
import pandas as pd
import geopandas as gpd
from tqdm.notebook import trange, tqdm

from paths import *
from functions import gdf_sp

<br>

## GAEMA

# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    os.path.join(output_path_tabs, 'tab_municipio_gaema.csv'),
)

# Results
df.info()
df.head()

# Merge
gdf_gaema = gdf_sp.merge(
    df,
    on='id_municipio',
    how='right',
)

# Results
gdf_gaema.head()

# Save geojson
gdf_gaema.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_gaema.geojson'),
    driver='GeoJSON',
    encoding='utf-8'
)

gdf_gaema.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_gaema.gpkg'),
    layer='sp_mpsp_gaema',
    driver='GPKG',
)

gdf_gaema.to_file(
    os.path.join(output_path_geo, 'sp_mpsp.gpkg'),
    layer='gaema',
    driver='GPKG',
)

# Results
gdf_gaema.head()

<br>

## Rede Protetiva

# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    os.path.join(output_path_tabs, 'tab_municipio_rp.csv'),
)

# Merge
gdf_rp = gdf_sp.merge(
    df,
    on='id_municipio',
    how='right',
)

# Save geojson
gdf_rp.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_rp.geojson'),
    driver='GeoJSON',
    encoding='utf-8'
)

gdf_rp.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_rp.gpkg'),
    layer='sp_mpsp_rp',
    driver='GPKG',
)

gdf_rp.to_file(
    os.path.join(output_path_geo, 'sp_mpsp.gpkg'),
    layer='rede_protetiva',
    driver='GPKG',
)

# Results
gdf_rp.head()

<br>

## Área Regional

# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    os.path.join(output_path_tabs, 'tab_municipio_ar.csv'),
)

# Merge
gdf_ar = gdf_sp.merge(
    df,
    on='id_municipio',
    how='left',
)

# Save geojson
gdf_ar.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_ar.geojson'),
    driver='GeoJSON',
    encoding='utf-8'
)

gdf_ar.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_ar.gpkg'),
    layer='sp_mpsp_ar',
    driver='GPKG',
)

gdf_ar.to_file(
    os.path.join(output_path_geo, 'sp_mpsp.gpkg'),
    layer='area_regional',
    driver='GPKG',
)

# Results
gdf_ar.head()

<br>

## Macroregiões

# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    os.path.join(output_path_tabs, 'tab_municipio_macroregioes.csv'),
)

# Merge
gdf_macroregioes = gdf_sp.merge(
    df,
    on='id_municipio',
    how='left',
)

# Save geojson
gdf_macroregioes.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_macroregioes.geojson'),
    driver='GeoJSON',
    encoding='utf-8'
)

gdf_macroregioes.to_file(
    os.path.join(output_path_geo, 'sp_mpsp_macroregioes.gpkg'),
    layer='sp_mpsp_macroregioes',
    driver='GPKG',
)

gdf_macroregioes.to_file(
    os.path.join(output_path_geo, 'sp_mpsp.gpkg'),
    layer='macroregioes',
    driver='GPKG',
)

# Results
gdf_macroregioes.head()
